requirements.txt
pytest == 8.3.4
requests == 2.32.3
pytest-mock == 3.14.0
psycopg2-binary == 2.9.10